requirements.txt
streamlit
google-generativeai
SpeechRecognition
gTTS
googletrans==4.0.0-rc1

